Expose additional citation fields in summary
The portal citation-feature needs additional citation-related attributes
available, such as usage limitations and licensing information.  Some of
these are already exposed, but all under the generic Constraints
tagname.  We repeat these here so as not to break current consumers of
that tag, using distinct tag names each.

This will also require an update to the schema plugins.